Renamed Diagnostic to Debug and deleted ItemSource (already defined)
